Extract the Alchemy getNFTs call to lib/alchemy
