+++ background/lib/alchemy.ts
@@ -289,3 +289,51 @@ export function transactionFromAlchemyWebsocketTransaction(
     network,
   }
 }
+
+// TODO move this someplace that isn't Alchemy-specific
+export type NFTItem = {
+  error?: string
+  media: { gateway?: string }[]
+  id: {
+    tokenId: string
+  }
+  contract: { address: string }
+  title: string
+  chainID: number
+}
+
+/**
+ * Use Alchemy's getNFTs call to get a wallet's NFT holdings across collections.
+ *
+ * Note that pagination isn't supported in this wrapper, so any responses after
+ * 100 NFTs will be dropped.
+ *
+ * More information https://docs.alchemy.com/reference/getnfts
+ *
+ * @param addressOnNetwork the address whose NFT portfolio we're fetching and
+ *        the network it should happen on.
+ */
+export async function getNFTs({
+  address,
+  network,
+}: AddressOnNetwork): Promise<NFTItem[]> {
+  // Today, only Polygon and Ethereum are supported
+  if (!["Polygon", "Ethereum"].includes(network.name)) {
+    return []
+  }
+
+  const requestUrl = new URL(
+    `https://${
+      network.name === "Polygon" ? "polygon-mainnet.g" : "eth-mainnet"
+    }.alchemyapi.io/nft/v2/${process.env.ALCHEMY_KEY}/getNFTs/`
+  )
+  requestUrl.searchParams.set("owner", address)
+  requestUrl.searchParams.set("filters[]", "SPAM")
+
+  // TODO validate data with ajv
+  const result = await (await fetch(requestUrl.toString())).json()
+  return result.ownedNfts.filter(
+    // TODO this is a misleading typeguard for unvalidated JSON
+    (nft: NFTItem) => typeof nft.error === "undefined"
+  )
+}
